// src/app/dashboard/alumnos/models/AlumnoModel.ts
export class AlumnoModel {
  id_alumno: string = '';
  nombre_alumno: string = '';
  apellido_p: string = '';
  apellido_m: string = '';
  telefono_1: string = '';
  telefono_2: string = '';
  curso_id: string = '';
}
export class VistaCursosAlm {
  id_curso: string = '';
  nombre_curso: string = '';
  descripsion: string = '';
  hora_inicio: string = '';
  hora_fin: string = '';
  id_alumno: string = '';
  nombre_alumno: string = '';
  apellido_p: string = '';
  apellido_m: string = '';
  telefono_1: string = '';
  telefono_2: string = '';
}